+++ usvg/src/tree/attributes.rs
@@ -672,18 +672,6 @@ pub enum FeImageKind {
 }
 
 
-/// A baseline shift value.
-#[allow(missing_docs)]
-#[derive(Clone, Copy, Debug)]
-pub enum BaselineShift {
-    Baseline,
-    Subscript,
-    Superscript,
-    Percent(f64),
-    Number(f64),
-}
-
-
 /// A path marker properties.
 #[derive(Clone, Debug)]
 pub struct PathMarker {